Set up the basics for the big first commit
Hopefully from this point, furture commits will be smaller, just
needed to get a good base going. Script should be capable of
downloading a config file in it's current state, so it's a good
first start. It can also pick apart short string for GitHub
references to pick apart a lot of information for the user.