Fix for Bash's inability to export array vars
Apparently using export doesn't work on arrays. This pretty much
invalidates teh work to export the arrays in the config, however
I don't see any reason to remove the exports, it might be handy for
something else later on. But, for now, the arrays will have to be
passed by hand.