Handle non-200 statuses better with curl command
Curl still succeeds and download data even if the HTTP status is not
a 200 (for example, a 404). In our use case, this is not a desired
state, so we need to check to make sure we downloaded the file we
were actually trying to download, and not an error page/message.